requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
praw==7.7.1
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
